Almost finished lesson 2.

+++ cplusplus/02_variables_and_types/05_declaration_and_definition/declaration.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-extern int pi;
+extern double pi;
 
 int main() {
