+++ cplusplus/02_variables_and_types/05_declaration_and_definition/definition.cpp
@@ -1,1 +1,2 @@
-int pi = 3.1415926536;
+// Note that the "extern" keyword is NOT required.
+double pi = 3.1415926536;
